Add currency table management functions
This commit introduces two static member functions in the `CurrencyExtensions` module: `isCurrencyTableEmpty` and `insertDefaultValues`. The `isCurrencyTableEmpty` function checks if the currency table is empty by executing a SQL count command. The `insertDefaultValues` function inserts default currency values into the table if it is found to be empty. Additionally, a new SQL query `getCounted` is defined in the `CurrencyQuery` module to support counting entries in the Currency table.