Add workflow to delete branches after PR merge
This commit introduces a new GitHub Actions workflow named `delete_after_merge.yml` that automates the deletion of branches once a pull request is merged. The workflow is triggered upon closing a pull request and verifies if it was merged before using the `peter-evans/delete-branch` action to remove the corresponding branch. The new workflow file is also included in the solution items of the `Binnaculum.sln` file.