Add getBetweenDates method for trade retrieval
Introduces a new method `getBetweenDates` in the `Do` type of `TradeExtensions.fs` to retrieve trades between specified dates. This method constructs and executes a SQL command with parameters for the start and end dates. Updates the `TradesQuery` module in `TradesQuery.fs` to include the corresponding SQL query for fetching trades based on the date range.