Comment out BankAccountGesture Tapped event handler
The `Tapped` event handler for `BankAccountGesture` in
`BankAccountTemplate.xaml.cs` has been commented out. This
includes the event subscription and navigation logic to
`BankAccountPage`. A `TODO` comment has been added to indicate
that this functionality is temporarily disabled until there is
a clear plan for the information to display and how to present
it.

The functionality for navigating to `BankMovementCreator`
remains unchanged.