Comment out BrokerAccountGesture Tapped event handler
The `Tapped` event handler for `BrokerAccountGesture` in the
`BrokerAccountTemplate` class has been commented out. This
functionality, which navigated to `BrokerAcccountPage`, is
temporarily disabled due to pending decisions on the information
to display and its presentation.

A `TODO` comment has been added to indicate that this code
should be revisited and re-enabled once the necessary details
are finalized.